+++ gulpfile.js
@@ -1,6 +1,5 @@
 var gulp = require("gulp");
 var connect = require("gulp-connect");
-var api = require("gulp-static-api");
 var stylus = require("gulp-stylus");
 var nib = require("nib");
 var sopStyl = require("sop-styl");
@@ -21,31 +20,8 @@ var globs = {
 
 require('./tasks/news');
 require('./tasks/pages');
-
-function sortByDate(a, b) {
-  if (!b.attributes.date) {
-    return -1;
-  }
-  if (!a.attributes.date) {
-    return 1;
-  }
-
-  a = new Date(a.attributes.date);
-  b = new Date(b.attributes.date);
-
-  return (a > b) ?
-    -1 : (a < b) ?
-    1 : 0;
-}
-
-gulp.task("api", function () {
-  api({
-    glob: "src/news/*.md",
-    count: 2,
-    sortBy: sortByDate
-  })
-    .pipe(gulp.dest("./build/api/news"));
-});
+require('./tasks/work');
+require('./tasks/api');
 
 gulp.task("connect", function () {
   connect.server({
